Fix map updating bug.
The problem: Because callbacks were a set, the order in which they're
iterated over is non-deterministic. However, some callbacks are order
reliant. Specifically things like updating the in-memory terminal screen
before determining if we've changed levels before calling the
level-change detection logic. I changed the callbacks set in the proxies
to be lists instead to enforce order.

There was, at the same time, a problem with two brains existing and each
sending duplicate messages, which managed to confuse things so that
*sometimes* the map update code worked and *sometimes* it didn't. I
fixed that as well, though the brain should probably be made a singleton
or there should be a warning.

Also removed some old debugging code from the events dispatcher.