Fix mine detection bug.
Grah, friggin' mines. So the problem here is that I tried to optimize
mine detection (was this smart? probably not, I don't even know how long
the detection is taking... probably not that long). In doing that, I
stripped each line in the display of leading and trailing whitespace
which threw off the detection because indices no longer made any sense
and weren't correlated to one another.