bug 1: No colors on NAO
For some reason the telnet server on nethack.alt.org doesn't like strict
vt102 as a terminal type to display color. Switching to xterm-color
fixes the issue, but I'm still wondering if there's going to be some
problem with terminal emulation in the future.

I think while playing nethack, at least, things will be fine since I've
done that quite a bit locally and the escapes seem to be well supported
by vt102 at the moment.