Fix sokoban sometimes changing to mines
Until Brain is refactored into some more meaningful structure, there's
a couple as "mines" even though it's clearly not the mines.

Specifically: When in the upper levels of sokoban and traveling
downward. Mines obviously only exists off of "main" or "not
sure", it an never come out of "sokoban". Enforcing that here is the
easiest way to fix weird branching craziness.