Add a preloaded function for setting the PRNG
I'd like to be able to do some integration tests. Unfortunately it's
impossible to do so with nethack unless I can set the seed on the random
number generator so that I can get consistent worlds.

This is a swag at getting there. Right now it just seeds the PRNG with
the same number everytime, but I'd like to be able to see it with
something set on (perhaps) the environment eventually.